pyproject.toml
[project]
name = "nano-banana"
version = "0.1.0"
description = "Claude Code skill for generating images via the Gemini API"
requires-python = ">=3.11"
dependencies = [
    "google-genai",
    "Pillow",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "pytest",
]
